#1638 refactor: categories/tags abstract base class
